Add linear upper bound approximation for Praos leader election
The linear function `f*sigma` is a lower bound for the Praos leader election and
can be used to quickly decide whether someone can be a leader.

This adds a parallel linear upper bound to quickly check whether someone cannot
be a slot leader. Uses calculus to compute the distance of the parallel linear
function and basic trigonometry (arctan, sine) to calculate the y-axis
displacement.