Performance optimisation of 'likelihood'.
Per the micro-benchmark for likelihood, this reduces the time from
around 50-60ms to around 1. Since this accounts for about 60-70% of the
time spend in `createRUpd`, we should hope to see at least a doubling of
performance there.

The optimisation is achieved by calculating `Set.fromList` only once,
rather than for each reward calculation.