"Added Sparse serialisation to Cardano.Ledger.ShelleyMA.TxBody
We added sparse (one field at a time) serialiation for Cardano.Ledger.ShelleyMA.TxBody.
This was carefully designed so that is backward comaptible with Shelley.Spec.Ledger.TxBody.
Shelley.Spec.Ledger.TxBody should deserialise as Cardano.Ledger.ShelleyMA.TxBody.
We added roundtrip property tests for Cardano.Ledger.ShelleyMA.TxBody, and rountrip
embeddin tests for Shelley.Spec.Ledger.TxBody.

We added the module Data.CodersInline which is a separate library (interchangeable with
Data.Coders) which should compile to faster code as it enables inlineing. Some
benchmarks show that it is only marginally faster than Data.coders. Furthee investigation
shows that cross module inlining is not working. We leave that for another day. In single
module code it was considerably faster. No code uses the new module except the benchmarks.
To get the benchmarks to work we needed to add a few NFData instances."